+++ src/components/Proposals/Search/index.ts
@@ -1,3 +1,3 @@
-export { default as SchemaSearch } from './SchemaSearch';
+export { default as SchemeSearch } from './SchemeSearch';
 export { default as StatusSearch } from './StatusSearch';
 export { default as TitleSearch } from './TitleSearch';
