+++ src/hooks/useParams.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect , useMemo, ChangeEvent} from 'react';
+import { useState, useEffect ,  ChangeEvent} from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
