+++ src/pages/proposals/index.tsx
@@ -25,7 +25,7 @@ import { FiFeather, FiCheckCircle, FiCheckSquare } from 'react-icons/fi';
 import { useProposals } from 'hooks/useProposals';
 import {
   StatusSearch,
-  SchemaSearch,
+  SchemeSearch,
   TitleSearch,
 } from '../../components/Proposals/Search';
 import { ProposalsWrapper, SidebarWrapper, ProposalTableHeaderActions, NewProposalButton, LoadingBox, FooterWrap, TableProposal, UnstyledAnchor, StyledTableRow } from './styles';
@@ -56,7 +56,7 @@ const ProposalsPage = observer(() => {
           </NewProposalButton>
           <TitleSearch />
           <StatusSearch />
-          <SchemaSearch />
+          <SchemeSearch />
         </ProposalTableHeaderActions>
         <FooterWrap>
           <Footer />
